FileAPI/Blob-slice.html

<!doctype html>
<meta charset=utf-8>
<title>Blob slice</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/Blob.js"></script>
<div id="log"></div>
<script>
test(function() {
  var blob1, blob2;

  test_blob(function() {
    return blob1 = new Blob(["squiggle"]);
  }, {
    expected: "squiggle",
    type: "",
    desc: "blob1."
  });

  test_blob(function() {
    return blob2 = new Blob(["steak"], {type: "content/type"});
  }, {
    expected: "steak",
    type: "content/type",
    desc: "blob2."
  });

  var arrayBuffer = new ArrayBuffer(16);
  var int8View = new Int8Array(arrayBuffer);
  for (var i = 0; i < 16; i++) {
    int8View[i] = i + 65;
  }

  var blob3 = new Blob([
    (new Uint8Array([0, 255, 0])).buffer,
    new Blob(['abcd']),
    'efgh',
    'ijklmnopqrstuvwxyz'
  ]);

  var testData = [
    // Test 3 strings
    [
      ["foo", "bar", "baz"],
      {},
      [{start: 0, length: 9, contents: "foobarbaz"},
       {start: 0, length: 3, contents: "foo"},
       {start: 3, length:6, contents:  "barbaz"},
       {start: 6, length: 3, contents:  "baz"},
       {start: 6, length: 6, contents: "baz"},
       {start: 0, length: 9, contents:  "foobarbaz"},
       {start: 0, length: 11, contents: "foobarbaz"},
       {start: 10, length: 5, contents: ""}]
    ],

    // Test string, Blob, string
    [
      ["foo", blob1, "baz"],
      {},
      [{start: 0, length: 3, contents:  "foo"},
       {start: 3, length: 8, contents:  "squiggle"},
       {start: 2, length: 2, contents:  "os"},
       {start: 10, length: 2, contents: "eb"}]
    ],

    // Test blob, string, blob
    [
      [blob1, "foo", blob1],
      {},
      [{start: 0, length: 8, contents:  "squiggle"},
       {start: 7, length: 2, contents:  "ef"},
       {start: 10, length: 2, contents: "os"},
       {start: 1, length: 3, contents:  "qui"},
       {start: 12, length: 3, contents: "qui"},
       {start: 40, length: 20, contents: ""}]
    ],

    // Test blobs all the way down
    [
      [blob2, blob1, blob2],
      {},
      [{start: 0, length: 5, contents:  "steak"},
       {start: 5, length: 8, contents:  "squiggle"},
       {start: 13, length: 5, contents: "steak"},
       {start: 1, length: 2, contents:  "te"},
       {start: 6, length: 4, contents:  "quig"}]
    ],

    // Test an ArrayBufferView
    [
      [int8View, blob1, "foo"],
      {},
      [{start: 0, length: 8, contents:  "ABCDEFGH"},
       {start: 8, length:10, contents:  "IJKLMNOPsq"},
       {start: 17, length: 3, contents: "qui"},
       {start: 4, length: 8, contents:  "EFGHIJKL"}]
    ],

    // Test a partial ArrayBufferView
    [
      [new Uint8Array(arrayBuffer, 3, 5), blob1, "foo"],
      {},
      [{start: 0, length: 8, contents:  "DEFGHsqu"},
       {start: 8, length:10, contents:  "igglefoo"},
       {start: 4, length: 8, contents:  "Hsquiggl"}]
    ],

    // Test type coercion of a number
    [
      [3, int8View, "foo"],
      {},
      [{start: 0, length: 8, contents:  "3ABCDEFG"},
       {start: 8, length:10, contents:  "HIJKLMNOPf"},
       {start: 17, length: 4, contents: "foo"},
       {start: 4, length: 8, contents:  "DEFGHIJK"}]
    ],

    [
      [blob3],
      {},
      [{start: 1, length: 3, contents:  "\uFFFD\u0000a"},
       {start: 4, length: 4, contents:  "bcde"},
       {start: 8, length: 4, contents:  "fghi"},
       {start: 1, length: 11, contents:  "\uFFFD\u0000abcdefghi"}]
    ]
  ];

  testData.forEach(function(data, i) {
    var blobs = data[0], options = data[1], tests = data[2];
    tests.forEach(function(expectations, j) {
      test(function() {
        var blob = new Blob(blobs, options);
        assert_true(blob instanceof Blob);
        assert_false(blob instanceof File);

        test_blob(function() {
          return blob.slice(expectations.start, expectations.start + expectations.length);
        }, {
          expected: expectations.contents,
          type: "",
          desc: "Slicing test: slice (" + i + "," + j + ")."
        });
      }, "Slicing test (" + i + "," + j + ").");
    });
  });
}, "Slicing");
</script>
